save one `MVar` per `IVar`
The original reason for having both a lock field and a field holding
the value of the `IVar` was that there was no way to read an `MVar`
without temporarily emptying it; `readMVar` used to do `takeMVar`
followed by `putMVar`. This changed a while ago, and `readMVar` is
now atomic. This allows the `IVar` code to be simplified.